refactor: remove smallvec from AuthUnixParams gids()
Changes AuthUnixParams to accept any iterable source of u32 values, and
return a slice of u32 IDs, instead of exposing a smallvec buffer.

This is implemented using a fixed-capacity stack-allocated array with
variable number of values stored, up to the maximum of 16 allowed by the
protocol.

This is a breaking change.