test(fuzz): ignore padding bytes in roundtrip test
Padding bytes can be anything as far as this deserialiser is involved
(though the spec recommends using zeros). Previously if padding bytes in
the input were not zeros, the buffer equality assertion would fail as
zeros are used in the serialised output.

This changes the fuzz test to no longer be sensitive to padding bytes by
comparing the deserialised messages (after padding is discarded) rather
than the wire-serialised buffers.